home page update
